+++ integration/single_server_test.go
@@ -8,10 +8,13 @@ import (
 	"io/ioutil"
 	"net/http"
 	"os"
+	"path"
 	"strconv"
 	"strings"
+	"time"
 
 	influxdb "github.com/influxdb/influxdb/client"
+	"github.com/influxdb/influxdb/datastore"
 	. "github.com/influxdb/influxdb/integration/helpers"
 	. "launchpad.net/gocheck"
 )
@@ -141,6 +144,36 @@ func (self *SingleServerSuite) TestListSeriesRegex(c *C) {
 	c.Assert(s, HasLen, 1)
 	maps := ToMap(s[0])
 	c.Assert(maps, HasLen, 3)
+}
+
+func (self *SingleServerSuite) TestDeleteExpiredShards(c *C) {
+	db := "delete_expired_shards"
+	client := self.server.GetClient(db, c)
+	c.Assert(client.CreateDatabase(db), IsNil)
+	err := client.CreateShardSpace(db, &influxdb.ShardSpace{
+		Name:            "default",
+		Regex:           ".*",
+		RetentionPolicy: "7d",
+		ShardDuration:   "1y",
+	})
+	c.Assert(err, IsNil)
+
+	data := CreatePoints("test_using_deleted_shard", 1, 1000000)
+	data[0].Columns = append(data[0].Columns, "time")
+	for i := range data[0].Points {
+		data[0].Points[i] = append(data[0].Points[i], 0)
+	}
+	for i := 0; i < 2; i++ {
+		err = client.WriteSeriesWithTimePrecision(data, influxdb.Second)
+		c.Assert(err, IsNil)
+	}
+	// wait for the retention sweep to kick in
+	time.Sleep(self.server.RetentionSweepPeriod() + time.Second)
+	f, err := os.Open(path.Join(self.server.DataDir(), datastore.SHARD_DATABASE_DIR))
+	c.Assert(err, IsNil)
+	dirs, err := f.Readdirnames(0)
+	c.Assert(err, IsNil)
+	c.Assert(dirs, HasLen, 0)
 }
 
 func (self *SingleServerSuite) TestListSeriesWithSpace(c *C) {